divide system test into their respective files
